# packages/contract/openapi.yaml
openapi: 3.0.3
info:
  title: MeterBase API
  version: "1.0.0"
  description: >
    Usage metering & monetization infrastructure. Ingest usage events, define
    meters, query time-bucketed aggregates, compute cost from pricing models,
    and fire usage-threshold alerts via signed webhooks.

    This document is the source of truth for the v1 REST contract. JSON
    properties are camelCase; the implementation maps to snake_case at the DB
    boundary. All endpoints are scoped to the organization that owns the API key.
servers:
  - url: http://localhost:48888
    description: Local
  - url: https://api.meterbase.example
    description: Production (placeholder)
security:
  - ApiKeyAuth: []
...
paths:
...
  /v1/customers/{id}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [Customers]
      summary: Get a customer
      operationId: getCustomer
      responses:
        '200':
          description: The customer
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Customer' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Customers]
      summary: Update a customer
      operationId: updateCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CustomerPatch' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Customer' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/plans:
    post:
      tags: [Pricing]
      summary: Create a plan
      operationId: createPlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Plan' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/plans/{id}/rate-cards:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    post:
      tags: [Pricing]
      summary: Add a rate card to a plan
      operationId: createRateCard
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RateCardCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RateCard' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/plans/{id}/cost:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [Pricing]
      summary: Compute the cost of a customer's usage for a period
      description: >
        Computes cost from usage and the plan's rate cards. Does NOT issue an
        invoice or move money. Must be deterministic and reproducible from raw
        events.
      operationId: computeCost
      parameters:
        - { in: query, name: customerId, required: true, schema: { type: string } }
        - { in: query, name: from, required: true, schema: { type: string, format: date-time } }
        - { in: query, name: to, required: true, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: Computed cost
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CostResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: "Pass the API key as a bearer token. Authorization: Bearer <api_key>."
  parameters:
...
    Id:
      in: path
      name: id
      required: true
      schema: { type: string }
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    PricingModel:
      type: string
      enum: [PAYG, FLAT_PLUS_OVERAGE, TIERED]
...
    CustomerCreate:
      type: object
      required: [externalId]
      properties:
        externalId: { type: string, description: The org's own id for this customer. }
        name: { type: string, nullable: true }
        metadata:
          type: object
          additionalProperties: true
          default: {}
    CustomerPatch:
      type: object
      properties:
        name: { type: string, nullable: true }
        metadata:
          type: object
          additionalProperties: true
    Customer:
      allOf:
        - $ref: '#/components/schemas/CustomerCreate'
        - type: object
          required: [id, createdAt]
          properties:
            id: { type: string }
            createdAt: { type: string, format: date-time }
    Plan:
      type: object
      required: [id, name, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        createdAt: { type: string, format: date-time }
    RateCardCreate:
      type: object
      required: [meterId, model, config]
      properties:
        meterId: { type: string }
        model: { $ref: '#/components/schemas/PricingModel' }
        currency: { type: string, default: USD }
        config:
          type: object
          additionalProperties: true
          description: >
            Model-specific config. PAYG: { unitPrice }. FLAT_PLUS_OVERAGE:
            { flatFee, included, overagePrice }. TIERED: { mode: "graduated"|"volume",
            tiers: [ { upTo, unitPrice } ] } (last tier upTo null = infinity).
          example:
            mode: graduated
            tiers:
              - { upTo: 1000, unitPrice: 0.01 }
              - { upTo: null, unitPrice: 0.005 }
    RateCard:
      allOf:
        - $ref: '#/components/schemas/RateCardCreate'
        - type: object
          required: [id, planId]
          properties:
            id: { type: string }
            planId: { type: string }
    CostResponse:
      type: object
      required: [currency, total, lineItems]
      properties:
        currency: { type: string }
        total: { type: number }
        lineItems:
          type: array
          items:
            type: object
            required: [meter, usage, cost]
            properties:
              meter: { type: string }
              usage: { type: number }
              cost: { type: number }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }